requirements.txt
asgiref==3.2.3
bcrypt==3.1.7
cffi==1.13.2
Django==3.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
psycopg2==2.8.4
pycparser==2.19
PyJWT==1.7.1
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
